[AT-1243] Fix repolinter on GitHub machines (#112)
### Summary

<!--- General summary / title -->

The Repolinter step works on self-hosted runners but fails during cron jobs running on GitHub hosted machines. This PR fixes those failures.

### Description

<!--- Details of what you changed -->

Now the repolinter ruleset file should successfully download on Windows using the aws cli; previously the region was incorrect or was missing. On mac `awscli` already existed and was seemingly not installed using homebrew and `brew install` failed with a failure to link; now it tries to install only if a version doesn't already exist.

### Related Issue

<!--- Link to issue where this is tracked -->

[Open Source Review - Run & Fix Repolinter](https://bitquill.atlassian.net/browse/AT-1243).