[AT-1270] SQLColumns handle empty input (#110)
### Summary

<!--- General summary / title -->

Make SQLColumns handle empty and null catalog and schema names in accordance with official documentation.

### Description

<!--- Details of what you changed -->

#### `SQL_ATTR_METADATA_ID` is `SQL_FALSE` (default)
`catalogName`, `schemaName`, and `tableName` are treated as search patterns. If `catalogName`, `schemaName`, `tableName`, or `columnName` are null they are replaced with `"%"`.

#### `SQL_ATTR_METADATA_ID` is `SQL_TRUE`
`catalogName`, `schemaName`, and `tableName` are treated as identifiers. If database is reported as a schema an error will be thrown if `schemaName` is null, likewise if the database is reported as a catalog.

Regardless of configuration, if both `schemaName` and `catalogName` are empty strings then `SQL_SUCCESS_WITH_INFO` is returned and a message is sent with a general warning.

Four tests have been added, for each combination of a database reported as catalog or schema and `SQL_ATTR_METADATA_ID` being true or false.

### Related Issue

<!--- Link to issue where this is tracked -->

[Fix implementation for SQLColumns](https://bitquill.atlassian.net/browse/AT-1270).