[AT-1299] Fix unexpected output for GetOdbcErrorMessage (#126)
### Summary

<!--- General summary / title -->

Fix unexpected output for `GetOdbcErrorMessage`

### Description

<!--- Details of what you changed -->

When `SQL_NO_DATA` is returned the message returned by `GetOdbcErrorMessage` on Linux is unpredictable as calls to `SQLGetDiagRec` can be intercepted by the driver manager and routed to the driver's `SQLGetDiagRecW` method and leave the message buffer uninitialized. This means the buffer is left full of garbage values that may or may not contain sensitive information.

Now the message in `GetOdbcErrorMessage` defaults to an empty array instead of an array full of garbage values so that if the driver manager leaves the message buffer empty this can be detected and "Cannot find ODBC error message" will be returned.

This fix only works when the driver manager leaves the buffer empty, so it will always be consistent on Linux. On Windows a return value of `SQL_NO_DATA` may or may not mean there is a message that can be read. Regardless, there won't be any sensitive data printed or garbage values. 

### Related Issue

<!--- Link to issue where this is tracked -->

[SQLGetDiagRec reads unexpected info (Linux)](https://bitquill.atlassian.net/browse/AT-1299).

### Additional Reviewers
@alexey-temnikov
@alinaliBQ
@RoyZhang2022
@trevorbonas
<!-- Any additional reviewers -->