[AT-1288] Port old performance test tool to new repository (#138)
### Summary

<!--- General summary / title -->

[AT-1288] Port old performance test tool to new repository.

### Description

<!--- Details of what you changed -->

The old performance test tool as it is in this PR (https://github.com/Bit-Quill/timestream-odbc-deprecated/pull/217) has been ported to the new driver repository and placed in its own `performance` directory, outside `src`. 

This performance tool does not rely on any of our source code and uses `sql.h` and `sqlext.h` for ODBC function definitions; for testing it just uses a connection string. The connection string is set in `performance/ITODBCHelper/it_odbc_helper.cpp` and defaults to `"DSN=timestream-iam"`.

Performance tests only print to stdout, results are no longer written to any files, that will need to be done in a separate ticket.

### Related Issue

<!--- Link to issue where this is tracked -->

[Migrate old driver performance tool into the new repository](https://bitquill.atlassian.net/browse/AT-1288).